requirements.txt
cortexutils
numpy
Pillow
